requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
python-vlc==3.0.21203
pycaw==20240210
comtypes==1.4.8
